+++ lib/src/gg_localize_refs.dart
@@ -8,6 +8,9 @@ import 'package:args/command_runner.dart';
 import 'package:gg_localize_refs/src/commands/unlocalize_refs.dart';
 import 'commands/localize_refs.dart';
 import 'package:gg_log/gg_log.dart';
+import 'package:gg_localize_refs/src/commands/get_ref_version.dart';
+import 'package:gg_localize_refs/src/commands/set_ref_version.dart';
+import 'package:gg_localize_refs/src/commands/get_version.dart';
 
 /// The command line interface for GgToLocal
 class GgToLocal extends Command<dynamic> {
@@ -15,6 +18,9 @@ class GgToLocal extends Command<dynamic> {
   GgToLocal({required this.ggLog}) {
     addSubcommand(LocalizeRefs(ggLog: ggLog));
     addSubcommand(UnlocalizeRefs(ggLog: ggLog));
+    addSubcommand(GetRefVersion(ggLog: ggLog));
+    addSubcommand(SetRefVersion(ggLog: ggLog));
+    addSubcommand(GetVersion(ggLog: ggLog));
   }
 
   /// The log function