+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Initial boilerplate.
 - Add git parameter to localize refs get
 - Add commands to export
+- Add tests for get ref and set ref
+- Add get version command
+- Add tests for process dependencies
+- add .idea to .gitignore
 - Add publish\_to: none when localizing
 
 ### Changed